convert workers page
