--- web/src/pages/workers/ViewWorkersPage.js
@@ -1,202 +1,0 @@
-import React, { Component } from 'react'
-import { Link } from "react-router-dom"
-
-export default class ViewWorkersPage extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      "loading": true,
-      "workers": [],
-      "executors": []
-    };
-  }
-
-  componentDidMount() {
-    Promise.all([
-      this.props.apiClient.getWorkers(),
-      this.props.apiClient.getExecutors()
-    ])
-      .then(([workers, executors]) => this.setState({workers, executors}))
-      .finally(() => this.setState({loading: false}))
-  }
-
-  onReplicate = (moduleName, args, intervalSeconds) => {
-    this.props.redirectTo(
-      `/workers/create?module_name=${moduleName}&args=${encodeURIComponent(JSON.stringify(args))}&interval_seconds=${intervalSeconds}`
-    )
-  }
-
-  onRemove = workerId => {
-    if (window.confirm(`Are you sure you want to remove worker ${workerId}`)) {
-      this.props.apiClient.removeWorker(workerId)
-        .then(() => {
-          this.setState({
-            ...this.state,
-            "workers": this.state.workers.filter(worker => worker["worker_id"] !== workerId),
-          });
-          this.props.showOkMessage(`Removed worker ${workerId}`)
-        })
-        .catch(error => {
-          console.error(error.response.data);
-          this.props.showErrorMessage(JSON.stringify(error.response.data))
-        });
-    }
-  }
-
-  onUpdateIntervalSeconds = (workerId, intervalSeconds) => {
-    this.props.apiClient.updateWorkerIntervalSeconds(workerId, intervalSeconds)
-      .then(() => {
-        this.props.showOkMessage(`Updated worker ${workerId} with interval seconds ${intervalSeconds}`)
-      })
-      .catch(error => {
-        console.error(error.response.data);
-        this.props.showErrorMessage(JSON.stringify(error.response.data))
-      });
-  }
-
-  onUpdateErrorResiliency = (workerId, errorResiliency) => {
-    this.props.apiClient.updateWorkerErrorResiliency(workerId, errorResiliency)
-      .then(() => {
-        this.props.showOkMessage(`Updated worker ${workerId} with error resiliency ${errorResiliency}`)
-      })
-      .catch(error => {
-        console.error(error.response.data);
-        this.props.showErrorMessage(JSON.stringify(error.response.data))
-      });
-  }
-
-  onUpdateExecutor = (workerId, executor) => {
-    this.props.apiClient.updateWorkerExecutor(workerId, executor)
-      .then(() => {
-        this.props.showOkMessage(`Updated worker ${workerId} with executor ${executor}`)
-      })
-      .catch(error => {
-        console.error(error.response.data);
-        this.props.showErrorMessage(JSON.stringify(error.response.data))
-      });
-  }
-
-  renderWorkersTable(workers) {
-    const nowSeconds = Math.floor(Date.now() / 1000)
-    return (
-      <table>
-        <thead>
-        <tr>
-          <th>ID</th>
-          <th>Interval (seconds)</th>
-          <th>Error Resiliency (non-positive is none)</th>
-          <th>Executor</th>
-          <th>Executed minutes ago</th>
-          <th>Operations</th>
-        </tr>
-        </thead>
-        <tbody>
-        {
-          workers.map(worker => {
-            const {
-              "worker_id": workerId,
-              "module_name": moduleName,
-              args,
-              "interval_seconds": intervalSeconds,
-              "error_resiliency": errorResiliency,
-              "executor_slug": executor,
-              "last_executed_seconds": lastExecutedSeconds
-            } = worker;
-            return (
-              <tr key={workerId}>
-                <td>
-                  <Link to={`/worker/${workerId}`}>{workerId}</Link>
-                </td>
-                <td>
-                  <input type="number" value={intervalSeconds} onChange={e => {
-                    e.preventDefault();
-                    this.setState({
-                      "workers": this.state.workers.map(worker => {
-                        if (worker["worker_id"] !== workerId) {
-                          return worker
-                        }
-                        return {
-                          ...worker,
-                          "interval_seconds": parseInt(e.target.value)
-                        }
-                      })
-                    })
-                  }}/>
-                  <button onClick={e => {
-                    e.preventDefault()
-                    this.onUpdateIntervalSeconds(workerId, intervalSeconds)}
-                  }>Update</button>
-                </td>
-                <td>
-                  <input type="number" min={-1} value={errorResiliency} onChange={e => {
-                    e.preventDefault();
-                    this.setState({
-                      "workers": this.state.workers.map(worker => {
-                        if (worker["worker_id"] !== workerId) {
-                          return worker
-                        }
-                        return {
-                          ...worker,
-                          "error_resiliency": parseInt(e.target.value)
-                        }
-                      })
-                    })
-                  }}/>
-                  <button onClick={e => {
-                    e.preventDefault()
-                    this.onUpdateErrorResiliency(workerId, errorResiliency)}
-                  }>Update</button>
-                </td>
-                <td>
-                  <select value={executor} onChange={e => {
-                    e.preventDefault();
-                    this.setState({
-                      "workers": this.state.workers.map(worker => {
-                        if (worker["worker_id"] !== workerId) {
-                          return worker
-                        }
-                        return {
-                          ...worker,
-                          "executor_slug": e.target.value
-                        }
-                      })
-                    })
-                  }}>
-                    {this.state.executors.map((executor, i) =>
-                      <option key={i} value={executor}>{executor}</option>
-                    )}
-                  </select>
-                  <button onClick={e => {
-                    e.preventDefault()
-                    this.onUpdateExecutor(workerId, executor)}
-                  }>Update</button>
-                </td>
-                <td>{lastExecutedSeconds !== -1 ? Math.floor((nowSeconds - lastExecutedSeconds) / 60) : 'N/A'}</td>
-                <td>
-                  <button onClick={e => {
-                    e.preventDefault();
-                    this.onReplicate(moduleName, args, intervalSeconds)
-                  }}>Create from</button>
-                  <button onClick={e => {
-                    e.preventDefault();
-                    this.onRemove(workerId)
-                  }}>x</button>
-                </td>
-              </tr>
-            )
-          })
-        }
-        </tbody>
-      </table>
-    )
-  }
-
-  render() {
-    return (
-      <div>
-        <b>View workers</b>
-        { this.state.loading ? <div>Loading workers...</div> : this.renderWorkersTable(this.state["workers"]) }
-      </div>
-    )
-  }
-}
